+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Periodic celery task to queue pending tasks
+
 ## [0.27.0] 2022-08-23
 
 ### Added